Fix makefile to not run when Vagrant box and Docker version match Added provider attribute to packer file for uploading box

+++ Makefile
@@ -3,6 +3,7 @@ CURRENT_DOCKER_VERSION := $(shell curl -s https://docs.docker.com/engine/release
 CURRENT_DEBIAN_VERSION := $(shell curl -s https://cdimage.debian.org/debian-cd/current/amd64/iso-cd/ | grep -oE "debian-(.*)-amd64-netinst.iso" | sed -e 's/<[^>]*>//g' | cut -d">" -f 1 | sed 's/"//g' | head -n 1 | cut -d- -f2)
 OUTPUT_DIRECTORY := /tmp
 DATETIME := $(shell date "+%Y-%m-%d %H:%M:%S")
+PROVIDER := "virtualbox"
 
 .PHONY: all version build test clean_test upload clean
 
@@ -13,6 +14,7 @@ version:
 	@echo Current Docker Version: $(CURRENT_DOCKER_VERSION)
 	@echo Current Box Version: $(CURRENT_BOX_VERSION)
 	@echo Current Debian Version: $(CURRENT_DEBIAN_VERSION)
+	@echo Provider: $(DEFAULT_PROVIDER)
 	@echo "========================="
 	@echo ""
 
@@ -20,23 +22,39 @@ build:
 ifeq ($(CURRENT_DOCKER_VERSION),$(CURRENT_BOX_VERSION))
 	@echo Nothing to do
 else
-	cd packer; packer build -var "docker_version=$(CURRENT_DOCKER_VERSION)" -var "debian_version=$(CURRENT_DEBIAN_VERSION)" -var "output_directory=/tmp" -only builder.virtualbox-iso.docker packer.pkr.hcl
+	cd packer; packer build -var "docker_version=$(CURRENT_DOCKER_VERSION)" -var "debian_version=$(CURRENT_DEBIAN_VERSION)" -var "output_directory=/tmp" -only builder.$(PROVIDER)-iso.docker packer.pkr.hcl
 endif
 
 test:
-	vagrant box add --name "testing-docker-box" $(OUTPUT_DIRECTORY)/packer-build/output/boxes/docker/$(CURRENT_DOCKER_VERSION)/virtualbox/docker.box
+ifeq ($(CURRENT_DOCKER_VERSION),$(CURRENT_BOX_VERSION))
+	@echo Nothing to do
+else
+	vagrant box add --name "testing-docker-box" $(OUTPUT_DIRECTORY)/packer-build/output/boxes/docker/$(CURRENT_DOCKER_VERSION)/$(PROVIDER)/docker.box
 	mkdir -p $(OUTPUT_DIRECTORY)/vagrant-docker-test; cd $(OUTPUT_DIRECTORY)/vagrant-docker-test; vagrant init testing-docker-box; \
-	vagrant up --provider virtualbox; \
+	vagrant up --provider $(PROVIDER); \
 	vagrant provision; \
 	DOCKER_HOST="tcp://$(vagrant ssh-config | grep -i "HostName" | awk '{ print $2 }'):$(vagrant port --guest 2375)/" $(HOME)/.Yohnah/Docker/docker run hello-world; \
 	vagrant destroy -f 
+endif
 
 clean_test:
+ifeq ($(CURRENT_DOCKER_VERSION),$(CURRENT_BOX_VERSION))
+	@echo Nothing to do
+else
 	vagrant box remove testing-docker-box || true
 	rm -fr $(OUTPUT_DIRECTORY)/vagrant-docker-test
+endif
 
 upload:
-	cd Packer; packer build -var "input_directory=$(OUTPUT_DIRECTORY)" -var "version=$(NEW_VERSION)" -var "version_description=Build version: $(DATETIME)" upload-box-to-vagrant-cloud.pkr.hcl
+ifeq ($(CURRENT_DOCKER_VERSION),$(CURRENT_BOX_VERSION))
+	@echo Nothing to do
+else
+	cd Packer; packer build -var "input_directory=$(OUTPUT_DIRECTORY)" -var "version=$(NEW_VERSION)" -var "version_description=Build version: $(DATETIME)" -var "provider=$(PROVIDER)" upload-box-to-vagrant-cloud.pkr.hcl
+endif
 
 clean: clean_test
+ifeq ($(CURRENT_DOCKER_VERSION),$(CURRENT_BOX_VERSION))
+	@echo Nothing to do
+else
 	rm -fr $(OUTPUT_DIRECTORY)/packer-build
+endif
